useDebugValue usage
