+++ src/templates/Home/index.jsx
@@ -1,70 +1,41 @@
-import P from "prop-types";
-import React, {
-  useState,
-  useLayoutEffect,
-  useRef,
-  forwardRef,
-  useImperativeHandle,
-} from "react";
+import React, { useDebugValue } from "react";
 
-export const Home = () => {
-  const [counted, setCounted] = useState([0, 1, 2, 3, 4]);
+const useMediaQuery = (queryValue, initialValue = false) => {
+  const [match, setMatch] = React.useState(initialValue);
 
-  const divRef = useRef(counted);
+  useDebugValue(`Query: ${queryValue}`, (name) => {
+    return `Media query: ${name}`;
+  });
 
-  useLayoutEffect(() => {
-    const now = Date.now();
-    while (Date.now() < now + 300);
-    console.log(divRef.current.otherDivRef.scrollHeight);
-    divRef.current.otherDivRef.scrollTop =
-      divRef.current.otherDivRef.scrollHeight;
-  }, [counted]);
+  React.useEffect(() => {
+    let isMounted = true;
+    const matchMedia = window.matchMedia(queryValue);
+    const handleChange = () => {
+      if (!isMounted) {
+        return;
+      }
 
-  const handleClick = () => {
-    setCounted((c) => [...c, c.slice(-1)[0] + 1]);
-    divRef.current.handleOtherClick();
-  };
+      setMatch(Boolean(matchMedia.matches));
+    };
 
-  return (
-    <>
-      <button onClick={handleClick}>Count {counted?.slice(-1)}</button>
-      <DisplayCounted counted={counted} ref={divRef} />
-    </>
-  );
+    matchMedia.addEventListener("change", handleChange);
+    setMatch(Boolean(matchMedia.matches));
+
+    return () => {
+      isMounted = false;
+      matchMedia.removeEventListener("change", handleChange);
+    };
+  }, [queryValue]);
+
+  return match;
 };
 
-export const DisplayCounted = forwardRef(function DisplayCounted(
-  { counted },
-  divRef,
-) {
-  const [rand, setRand] = useState("0.24");
-  const otherDivRef = useRef();
-
-  const handleOtherClick = () => {
-    setRand(Math.random().toFixed(2));
-  };
-
-  useImperativeHandle(divRef, () => ({
-    handleOtherClick,
-    otherDivRef: otherDivRef.current,
-  }));
-
+export const Home = () => {
+  const huge = useMediaQuery("(min-width: 600px)");
+  const background = huge ? "green" : "yellow";
   return (
-    <div
-      ref={otherDivRef}
-      style={{ height: "100px", width: "100px", overflowY: "scroll" }}
-    >
-      {counted.map((c) => {
-        return (
-          <p onClick={handleOtherClick} key={`c-${c}`}>
-            {c} +++ {rand}
-          </p>
-        );
-      })}
+    <div style={{ fontSize: "60px", background }}>
+      <h1>Oi</h1>
     </div>
   );
-});
-
-DisplayCounted.propTypes = {
-  counted: P.array,
 };
